fix: clawback after slash fix, refactor test helper
