+++ x/auth/vesting/types/vesting_account_test.go
@@ -14,8 +14,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	"github.com/cosmos/cosmos-sdk/x/auth/vesting/types"
-	"github.com/cosmos/cosmos-sdk/x/staking"
-	stakingkeeper "github.com/cosmos/cosmos-sdk/x/staking/keeper"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 )
 
@@ -955,41 +953,18 @@ func TestComputeClawback(t *testing.T) {
 	require.Equal(t, *va, va2)
 }
 
-func createValidator(t *testing.T, ctx sdk.Context, app *simapp.SimApp, powers int64) (sdk.ValAddress, stakingtypes.Validator) {
-	valTokens := sdk.TokensFromConsensusPower(powers, sdk.DefaultPowerReduction)
-	addrs := simapp.AddTestAddrsIncremental(app, ctx, 1, valTokens)
-	valAddrs := simapp.ConvertAddrsToValAddrs(addrs)
-	pks := simapp.CreateTestPubKeys(1)
-	cdc := simapp.MakeTestEncodingConfig().Marshaler
-
-	app.StakingKeeper = stakingkeeper.NewKeeper(
-		cdc,
-		app.GetKey(stakingtypes.StoreKey),
-		app.AccountKeeper,
-		app.BankKeeper,
-		app.GetSubspace(stakingtypes.ModuleName),
-	)
-
-	val, err := stakingtypes.NewValidator(valAddrs[0], pks[0], stakingtypes.Description{})
-	require.NoError(t, err)
-
-	app.StakingKeeper.SetValidator(ctx, val)
-	require.NoError(t, app.StakingKeeper.SetValidatorByConsAddr(ctx, val))
-	app.StakingKeeper.SetNewValidatorByPowerIndex(ctx, val)
-
-	_, err = app.StakingKeeper.Delegate(ctx, addrs[0], valTokens, stakingtypes.Unbonded, val, true)
-	require.NoError(t, err)
-
-	_ = staking.EndBlocker(ctx, app.StakingKeeper)
-
-	return valAddrs[0], val
-}
-
 func TestClawback(t *testing.T) {
 	c := sdk.NewCoins
 	fee := func(x int64) sdk.Coin { return sdk.NewInt64Coin(feeDenom, x) }
 	stake := func(x int64) sdk.Coin { return sdk.NewInt64Coin(stakeDenom, x) }
 	now := tmtime.Now()
+
+	// set up simapp and validators
+	app := simapp.Setup(false)
+	ctx := app.BaseApp.NewContext(false, tmproto.Header{}).WithBlockTime((now))
+	valAddr, val := createValidator(t, ctx, app, 100)
+	require.Equal(t, "stake", app.StakingKeeper.BondDenom(ctx))
+
 	lockupPeriods := types.Periods{
 		{Length: int64(12 * 3600), Amount: c(fee(1000), stake(100))}, // noon
 	}
@@ -1004,14 +979,12 @@ func TestClawback(t *testing.T) {
 	bacc, origCoins := initBaseAccount()
 	_, _, funder := testdata.KeyTestPubAddr()
 	va := types.NewClawbackVestingAccount(bacc, funder, origCoins, now.Unix(), lockupPeriods, vestingPeriods)
+	// simulate 17stake lost to slashing
+	va.DelegatedVesting = c(stake(17))
 	addr := va.GetAddress()
+	app.AccountKeeper.SetAccount(ctx, va)
 
-	// set up simapp and validators
-	ctx := app.BaseApp.NewContext(false, tmproto.Header{}).WithBlockTime((now))
-	valAddr, val := createValidator(t, ctx, app, 100)
-	require.Equal(t, "stake", app.StakingKeeper.BondDenom(ctx))
-
-	// fund the vesting account with 17 stake lost to slashing
+	// fund the vesting account with 17 take lost to slashing
 	err := simapp.FundAccount(app.BankKeeper, ctx, addr, c(fee(1000), stake(83)))
 	require.NoError(t, err)
 	require.Equal(t, int64(1000), app.BankKeeper.GetBalance(ctx, addr, feeDenom).Amount.Int64())
@@ -1030,7 +1003,8 @@ func TestClawback(t *testing.T) {
 
 	// clawback the unvested funds (600fee, 50stake)
 	_, _, dest := testdata.KeyTestPubAddr()
-	err = va.Clawback(ctx, dest, app.AccountKeeper, app.BankKeeper, app.StakingKeeper)
+	va2 := app.AccountKeeper.GetAccount(ctx, addr).(*types.ClawbackVestingAccount)
+	err = va2.Clawback(ctx, dest, app.AccountKeeper, app.BankKeeper, app.StakingKeeper)
 	require.NoError(t, err)
 
 	// check vesting account
@@ -1226,6 +1200,8 @@ func TestContinuousVestingAccountMarshal(t *testing.T) {
 	baseVesting := types.NewBaseVestingAccount(baseAcc, coins, time.Now().Unix())
 	acc := types.NewContinuousVestingAccountRaw(baseVesting, baseVesting.EndTime)
 
+	app := simapp.Setup(false)
+
 	bz, err := app.AccountKeeper.MarshalAccount(acc)
 	require.NoError(t, err)
 
@@ -1242,6 +1218,8 @@ func TestContinuousVestingAccountMarshal(t *testing.T) {
 func TestPeriodicVestingAccountMarshal(t *testing.T) {
 	baseAcc, coins := initBaseAccount()
 	acc := types.NewPeriodicVestingAccount(baseAcc, coins, time.Now().Unix(), types.Periods{types.Period{3600, coins}})
+
+	app := simapp.Setup(false)
 
 	bz, err := app.AccountKeeper.MarshalAccount(acc)
 	require.NoError(t, err)
@@ -1260,6 +1238,8 @@ func TestDelayedVestingAccountMarshal(t *testing.T) {
 	baseAcc, coins := initBaseAccount()
 	acc := types.NewDelayedVestingAccount(baseAcc, coins, time.Now().Unix())
 
+	app := simapp.Setup(false)
+
 	bz, err := app.AccountKeeper.MarshalAccount(acc)
 	require.NoError(t, err)
 
@@ -1276,6 +1256,8 @@ func TestDelayedVestingAccountMarshal(t *testing.T) {
 func TestPermanentLockedAccountMarshal(t *testing.T) {
 	baseAcc, coins := initBaseAccount()
 	acc := types.NewPermanentLockedAccount(baseAcc, coins)
+
+	app := simapp.Setup(false)
 
 	bz, err := app.AccountKeeper.MarshalAccount(acc)
 	require.NoError(t, err)
@@ -1296,6 +1278,8 @@ func TestClawbackVestingAccountMarshal(t *testing.T) {
 	acc := types.NewClawbackVestingAccount(baseAcc, addr, coins, time.Now().Unix(),
 		types.Periods{types.Period{3600, coins}}, types.Periods{types.Period{3600, coins}})
 
+	app := simapp.Setup(false)
+
 	bz, err := app.AccountKeeper.MarshalAccount(acc)
 	require.NoError(t, err)
 
@@ -1307,6 +1291,22 @@ func TestClawbackVestingAccountMarshal(t *testing.T) {
 	// error on bad bytes
 	_, err = app.AccountKeeper.UnmarshalAccount(bz[:len(bz)/2])
 	require.Error(t, err)
+}
+
+func TestClawbackVestingAccountStore(t *testing.T) {
+	baseAcc, coins := initBaseAccount()
+	addr := sdk.AccAddress([]byte("the funder"))
+	acc := types.NewClawbackVestingAccount(baseAcc, addr, coins, time.Now().Unix(),
+		types.Periods{types.Period{3600, coins}}, types.Periods{types.Period{3600, coins}})
+
+	app := simapp.Setup(false)
+	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
+	createValidator(t, ctx, app, 100)
+
+	app.AccountKeeper.SetAccount(ctx, acc)
+	acc2 := app.AccountKeeper.GetAccount(ctx, acc.GetAddress())
+	require.IsType(t, &types.ClawbackVestingAccount{}, acc2)
+	require.Equal(t, acc.String(), acc2.String())
 }
 
 func initBaseAccount() (*authtypes.BaseAccount, sdk.Coins) {
